Refactor: organise code base
- Organised and grouped files into industry standard organisation formats
